Inventory and Item Merge Preperation
Getting the Inventory to match the master for easier merging
